let IGDB router accepts path argument
