+++ lib/v2/cas/maintainer.js
@@ -1,6 +1,6 @@
 module.exports = {
     '/cg/:caty?': ['nczitzk'],
-    '/genetics/xshd': ['panyq357'],
+    '/genetics/:path+': ['panyq357'],
     '/ia/yjs': ['shengmaosu'],
     '/iee/kydt': ['nczitzk'],
     '/is/:path+': ['Misaka13514'],
