+++ lib/v2/cas/radar.js
@@ -11,10 +11,10 @@ module.exports = {
         ],
         'www.genetics': [
             {
-                title: '遗传与生物学研究所 - 学术活动',
+                title: '遗传与发育生物学研究所',
                 docs: 'https://docs.rsshub.app/routes/university#zhong-guo-ke-xue-yuan',
-                source: ['/jixs/yg', '/'],
-                target: '/cas/genetics/xshd',
+                source: ['/'],
+                target: '/cas/genetics/:path+',
             },
         ],
         'www.ia': [
