+++ lib/v2/cas/router.js
@@ -1,6 +1,6 @@
 module.exports = (router) => {
     router.get('/cg/:caty?', require('./cg/index'));
-    router.get('/genetics/xshd', require('./genetics/xshd'));
+    router.get('/genetics/:path+', require('./genetics/index'));
     router.get('/ia/yjs', require('./ia/yjs'));
     router.get('/iee/kydt', require('./iee/kydt'));
     router.get('/is/:path+', require('./is/index'));
