+++ website/docs/routes/university.mdx
@@ -3524,9 +3524,19 @@ jsjxy.hbut.edu.cn 证书链不全，自建 RSSHub 可设置环境变量 NODE\_TL
   | xwdt2016/tzgg2016 | xwdt2016/kjdt2016 | kxcb2016/kpdt2016 |
 </Route>
 
-### 遗传与发育生物学研究所 学术活动 {#zhong-guo-ke-xue-yuan-yi-chuan-yu-fa-yu-sheng-wu-xue-yan-jiu-suo-xue-shu-huo-dong}
+### 遗传与发育生物学研究所 {#zhong-guo-ke-xue-yuan-yi-chuan-yu-fa-yu-sheng-wu-xue-yan-jiu-suo}
 
-<Route author="panyq357" example="/cas/genetics/xshd" path="/cas/genetics/xshd" radar="1" />
+<Route author="panyq357" example="/cas/genetics/jixs/yg" path="/cas/genetics/:path+" paramsDesc={['路径，可在 URL 找到']} radar="1">
+  路径参数示例：
+
+  | 路径                  | 栏目       |
+  | :-------------------- | :--------- |
+  | jixs/yg               | 学术预告   |
+  | dtxw/kyjz             | 科研进展   |
+  | edu/zsxx/ssszs_187556 | 硕士生招生 |
+  | edu/zsxx/bsszs_187557 | 博士生招生 |
+
+</Route>
 
 ## 中国科学院大学 {#zhong-guo-ke-xue-yuan-da-xue}
